fix(packages): removed render functions into mapper
