+++ src/pages/Energy/ui/pages/SelectPackages/PlanDisplay.tsx
@@ -1,9 +1,8 @@
 import styles from './Buy.module.scss';
-import checkmark from './images/checkmark.svg';
 import { features, plans } from '../../../types/type';
-import IconsNumber from 'src/components/IconsNumber/IconsNumber';
 import AdviserHoverWrapper from 'src/features/adviser/AdviserHoverWrapper/AdviserHoverWrapper';
 import { Display } from 'src/components';
+import { renderPlanFeature } from './planFeatureMapper';
 
 interface PlanDisplayProps {
   plan: (typeof plans)[0];
@@ -12,74 +11,6 @@ interface PlanDisplayProps {
 }
 
 function PlanDisplay({ plan, index, isSelected }: PlanDisplayProps) {
-  const renderPlanFeature = (featureIndex: number) => {
-    const featureRenderers = {
-      default: () => (
-        <div
-          className={
-            plan.features[featureIndex] ? styles.checkmark : styles.excluded
-          }
-        >
-          {plan.features[featureIndex] ? (
-            <img src={checkmark} alt="checkmark" className={styles.checkmark} />
-          ) : (
-            '❌'
-          )}
-        </div>
-      ),
-      2: () =>
-        plan.uploads > 0 ? (
-          <div className={styles.uploads}>~ {plan.uploads}</div>
-        ) : (
-          <div className={styles.excluded}>❌</div>
-        ),
-      3: () => (
-        <div className={styles.symbols}>
-          <span className={styles.symbolsText}>
-            {'> '}
-            <span className={styles.symbolsDigit}>{plan.symbols}</span>
-            {' symbols'}
-          </span>
-        </div>
-      ),
-      4: () =>
-        plan.fuel > 0 ? (
-          <div className={styles.fuel}>
-            <IconsNumber value={plan.fuel} type="hydrogen" isVertical={false} />
-          </div>
-        ) : (
-          <div className={styles.excluded}>❌</div>
-        ),
-      5: () =>
-        plan.energy > 0 ? (
-          <div className={styles.energy}>
-            <span className={styles.energyIcon}>
-              <IconsNumber
-                value={plan.energy}
-                type="energy"
-                isVertical={false}
-              />
-            </span>
-          </div>
-        ) : (
-          <div className={styles.excluded}>❌</div>
-        ),
-      6: () =>
-        plan.influence > 0 ? (
-          <div className={styles.influence}>
-            {plan.influence} <span className={styles.gray}>‰</span>
-          </div>
-        ) : (
-          <div className={styles.excluded}>❌</div>
-        ),
-    };
-
-    const renderer =
-      featureRenderers[featureIndex as keyof typeof featureRenderers] ||
-      featureRenderers.default;
-    return renderer();
-  };
-
   return (
     <div
       className={`${styles.displayWrapper} ${isSelected ? styles.active : ''} ${
@@ -109,9 +40,9 @@ function PlanDisplay({ plan, index, isSelected }: PlanDisplayProps) {
       >
         <AdviserHoverWrapper adviserContent={'test'}>
           <div className={styles.planContent}>
-            {features.map((_, index) => (
-              <div key={index} className={styles.row}>
-                {renderPlanFeature(index)}
+            {features.map((_, featureIndex) => (
+              <div key={featureIndex} className={styles.row}>
+                {renderPlanFeature(plan, featureIndex)}
               </div>
             ))}
             <div className={styles.row}>
